+++ packages/react/src/hooks/participant-hooks.ts
@@ -9,7 +9,7 @@ import {
   ParticipantMedia,
   observeParticipantMedia,
 } from '@livekit/components-core';
-import { useObservableState } from '../utils';
+import { useObservableState } from '../hooks/utiltity-hooks';
 import { useEnsureParticipant, useRoomContext } from '../context';
 
 export type ParticipantFilter = Parameters<Participant[]['filter']>['0'];
